requirements.txt
pymongo
python-dotenv
pillow